+++ src/mewpy/cobra/com/__init__.py
@@ -13,3 +13,7 @@
 
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+from .analysis import *
+from .similarity import *
+from .steadycom import *
